refactor: move productList file
