+++ src/layouts/productsCatalogues.jsx
@@ -1,16 +1,9 @@
 import React, { useState } from 'react';
 
-import ProductsList from '../components/ui/productsList';
+import ProductsList from '../components/ui/productList/productsList';
 import { useHistory, useParams } from 'react-router';
 import { SuppliersCatalogueModelProvider } from '../model/useSuppliersCatalogueModel.js';
 import SupplierSelectField from '../components/ui/supplierSelectField.jsx';
-
-
-// const SearchBarDropDownItemContent = ({allSuppliersItem}) => (
-//     <>
-//         {allSuppliersItem.supplier}<span className="badge rounded-pill bg-light text-dark">{allSuppliersItem.productsCount} sku</span>
-//     </> 
-// );
 
 const ProductsCatalogue = () => {
     const [supplierId, setSupplierId] = useState(useParams().supplierId);
